Improve menu accessibility (#44)
This commit adds 2 new attributes to the MenuItem component: 'disabled' and 'href'.
'href' is used for specifying links associated with menu items. This is required to
make menu links accessible and navigable using keyboard. Currently, 'disabled' is
also used only for accessibility purposes: if a MenuItem is not disabled, a tabindex
attribute is added to help with keyboard navigation.

Additionally, an alt text is added to the logo image.